google_auth_services_server/requirements.txt
google-auth-oauthlib==1.0.0
google-auth==2.22.0
google-api-python-client==2.95.0
fastapi==0.100.0
uvicorn==0.23.0
python-jose==3.3.0
python-dotenv==1.0.0
pydantic==2.1.1
httpx==0.24.1
python-multipart==0.0.6
prometheus-client==0.17.1
python-json-logger==2.0.7
sentry-sdk==1.28.1
bcrypt==4.0.1
passlib==1.7.4
email-validator==2.0.0.post2
aioredis==2.0.1
redis==4.6.0
structlog==23.1.0
python-json-logger==2.0.7
tenacity==8.2.2
google-api-python-client-stubs==1.16.0
google-auth-httplib2==0.1.1
aiohttp==3.8.5 